Removed the visible bool prop from TimeTravel
