+++ src/constants/timeline.js
@@ -1,3 +1,2 @@
 
-export const TIMELINE_HEIGHT = '55px';
 export const MAX_TICK_SPACING_PX = 415;
